+++ app/Filters/ViewFilters.php
@@ -7,8 +7,6 @@ use Bayfront\Bones\Application\Services\Filters\FilterSubscription;
 use Bayfront\Bones\Application\Utilities\App;
 use Bayfront\Bones\Interfaces\FilterSubscriberInterface;
 use Bayfront\RouteIt\Router;
-use Bayfront\Translation\Translate;
-use Bayfront\Translation\TranslationException;
 
 /**
  * View-related filters.
@@ -17,16 +15,14 @@ class ViewFilters extends FilterSubscriber implements FilterSubscriberInterface
 {
 
     protected Router $router;
-    protected Translate $translate;
 
     /**
      * The container will resolve any dependencies.
      */
 
-    public function __construct(Router $router, Translate $translate)
+    public function __construct(Router $router)
     {
         $this->router = $router;
-        $this->translate = $translate;
     }
 
     /**
@@ -36,8 +32,7 @@ class ViewFilters extends FilterSubscriber implements FilterSubscriberInterface
     public function getSubscriptions(): array
     {
         return [
-            new FilterSubscription('webapp.response.data', [$this, 'addData'], 10),
-            new FilterSubscription('webapp.response.body', [$this, 'addTagSay'], 10)
+            new FilterSubscription('webapp.response.data', [$this, 'addData'], 10)
         ];
     }
 
@@ -56,52 +51,6 @@ class ViewFilters extends FilterSubscriber implements FilterSubscriberInterface
             ],
             'year' => date('Y')
         ]);
-    }
-
-    /**
-     * Add support for the @say: template tag which returns the translation of a given string.
-     *
-     * @param string $body
-     * @return string
-     */
-
-    public function addTagSay(string $body): string
-    {
-
-        // @say
-
-        preg_match_all("/@say:[\w.]+/", $body, $tags); // Any word character or period
-
-        if (isset($tags[0]) && is_array($tags[0])) { // If a tag was found
-
-            try {
-
-                foreach ($tags[0] as $tag) {
-
-                    $use = explode(':', $tag, 2);
-
-                    if (isset($use[1])) { // If valid @say syntax
-
-                        // Keep original string if not found
-
-                        $body = str_replace($tag, $this->translate->get($use[1], [], $use[1]), $body);
-
-                    }
-                }
-
-            } catch (TranslationException) {
-
-                /*
-                 * No translation exists for this tag.
-                 * Do nothing.
-                 */
-
-            }
-
-        }
-
-        return $body;
-
     }
 
 }