+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed Scheduler service
 - Removed `app.version` config value
 - Removed `@route:` template tag filter
+- Removed `@say:` template tag filter
 
 ## [3.1.0] - 2024.12.25